Separate build-server.ts responsabilities
