+++ src/server/build-server.ts
@@ -1,148 +1,23 @@
-import { compare, hash } from "bcrypt-ts";
-import fastify, {
-  FastifyInstance,
-  FastifyReply,
-  FastifyRequest,
-} from "fastify";
-import jwt from "jsonwebtoken";
-import { prisma } from "../setup-db";
-import { CustomError, errorHandler } from "./error-handler";
-import {
-  GetUserRequestBody,
-  PostAuthBodySchema,
-  PostUserBodySchema,
-  PostUserRequestBody,
-} from "./schemas";
-
-function isAuthenticated(request: FastifyRequest) {
-  const authHeader = request.headers.authorization;
-  if (!authHeader || !authHeader.startsWith("Bearer ")) {
-    throw new CustomError(
-      "Authentication failed. Log in, then try again",
-      "AUT_01",
-      "No authentication token of type Bearer was provided"
-    );
-  }
-  const token = authHeader.split(" ")[1];
-  const decoded = jwt.verify(token, process.env.TOKEN_KEY);
-  if (!decoded.id) {
-    throw new CustomError(
-      "Authentication failed. Try logging in once again",
-      "AUT_02",
-      "Decoded Payload from authentication token did not match the expected."
-    );
-  }
-}
+import fastify, { FastifyInstance } from "fastify";
+import { errorHandler } from "./error-handler";
+import { getUserRoute } from "./route-get-user";
+import { postAuthRoute } from "./route-post-auth";
+import { postUserRoute } from "./route-post-users";
+import { PostAuthBodySchema, PostUserBodySchema } from "./schemas";
 
 export function buildServer(): FastifyInstance {
   const app = fastify({});
-
   app.setErrorHandler(errorHandler);
 
   app.get("/hello", async () => {
     return { message: "Hello World!" };
   });
 
-  app.post(
-    "/users",
-    PostUserBodySchema,
-    async (
-      request: FastifyRequest<{ Body: PostUserRequestBody }>,
-      reply: FastifyReply
-    ) => {
-      isAuthenticated(request);
+  app.post("/users", PostUserBodySchema, postUserRoute);
 
-      const { id, name, email, password, birthDate } = request.body;
-      const hashedPassword = await hash(password, 8);
-      const newUser = await prisma.user.create({
-        data: {
-          name: name,
-          email: email,
-          password: hashedPassword,
-          birthDate: new Date(birthDate),
-        },
-      });
-      return reply.code(201).send({
-        id: newUser.id,
-        name: newUser.name,
-        email: newUser.email,
-        birthdate: newUser.birthDate,
-      });
-    }
-  );
+  app.post("/auth", PostAuthBodySchema, postAuthRoute);
 
-  app.post(
-    "/auth",
-    PostAuthBodySchema,
-    async (
-      request: FastifyRequest<{ Body: GetUserRequestBody }>,
-      reply: FastifyReply
-    ) => {
-      const { email, password, rememberMe } = request.body;
-      const user = await prisma.user.findUnique({
-        where: { email },
-      });
+  app.get("/users/:id", getUserRoute);
 
-      if (!user) {
-        throw new CustomError("Email not registered on platform", "EML_02");
-      }
-
-      const isPasswordValid = await compare(password, user.password);
-      if (!isPasswordValid) {
-        throw new CustomError("Wrong password. Try again", "PSW_03");
-      }
-
-      const token = jwt.sign({ id: user.id }, process.env.TOKEN_KEY, {
-        expiresIn: rememberMe
-          ? "1w"
-          : (Number(process.env.TOKEN_TIMEOUT) ?? 30),
-      });
-
-      return reply.code(200).send({
-        user: {
-          id: user.id,
-          name: user.name,
-          email: user.email,
-          birthDate: user.birthDate,
-        },
-        token: token,
-      });
-    }
-  );
-
-  app.get(
-    "/users/:id",
-    async (request: FastifyRequest<{ Params: { id: string } }>, reply) => {
-      isAuthenticated(request);
-      const { id } = request.params;
-
-      const userId = Number(id);
-      if (isNaN(userId)) {
-        throw new CustomError(
-          "Invalid ID. User ID must be a number",
-          "USR_02",
-          "The user ID must be an integer"
-        );
-      }
-
-      const user = await prisma.user.findUnique({
-        where: { id: Number(id) },
-      });
-
-      if (!user) {
-        throw new CustomError(
-          "User not found",
-          "USR_01",
-          "User id was not found on the database"
-        );
-      }
-      return reply.code(200).send({
-        id: user.id,
-        name: user.name,
-        email: user.email,
-        birthDate: user.birthDate,
-      });
-    }
-  );
   return app;
 }
